# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ExecuTorch Bencher API",
    "description": "Internal Android LLM benchmark collection and analysis service. Provides HTTP endpoints to ingest, store, and retrieve benchmark run results, content-addressed artifacts, and registered model assets.",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/results": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Group runs into benchmark configurations and report each one's\nmedian/min/max/n throughput over succeeded runs, plus failure and\nthrottling counts. Rows are ordered by commit timestamp (falling back to\nfirst run time) descending. At most 500 rows are returned.",
        "operationId": "getResults",
        "parameters": [
          {
            "name": "platform",
            "in": "query",
            "description": "One of the stable `Platform` values.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Platform"
            }
          },
          {
            "name": "device_class",
            "in": "query",
            "description": "One of the stable `DeviceClass` values.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/DeviceClass"
            }
          },
          {
            "name": "device_serial",
            "in": "query",
            "description": "Android: device serial. Linux: hostname.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "model_asset_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "git_commit_sha",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "git_branch",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "git_dirty",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sumd_driver_version",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bsp_version",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "host_accelerator",
            "in": "query",
            "description": "The accelerator as the runtime backend reports it.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "prompt_sha256",
            "in": "query",
            "description": "Lowercase 64-character hexadecimal SHA-256 digest of the prompt.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Grouped results and filter facets.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultsResponse"
                }
              }
            }
          },
          "400": {
            "description": "An invalid filter value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal storage error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiErrorBody": {
        "type": "object",
        "description": "The `error` object inside an [`ApiErrorResponse`].",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable, machine-readable failure category. Current values: `invalid_request`,\n`not_found`, `artifact_file_missing`, `payload_too_large`, `conflict`,\n`not_implemented`, `internal_error`. This field is an open string, not a\nclosed enum - new categories may be added as the API grows.",
            "example": "invalid_request"
          },
          "message": {
            "type": "string",
            "description": "Human-readable description of the failure. Not guaranteed stable across\nversions or releases - match on `code`, not this text.",
            "example": "gpu_clock_mhz must be greater than zero"
          },
          "details": {
            "description": "Optional structured context about the failure, e.g. which field was invalid."
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Opaque identifier correlating this error with server-side logs, when available."
          }
        }
      },
      "ApiErrorResponse": {
        "type": "object",
        "description": "The consistent JSON envelope every rejected or failed request responds with.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody"
          }
        }
      },
...
      "DeviceClass": {
        "type": "string",
        "description": "Whether a host is a lab device under full control or an external one.\nInternal devices (rooted development phones with SUMD/BSP/clock control)\nmust carry the rigorous Android snapshot; external hosts (retail phones,\nany Linux box) record whatever they can report. See\n`specs/benchmark-schema` - \"Host state is captured as a\nplatform-specific immutable snapshot\".",
        "enum": [
          "internal",
          "external"
        ]
      },
...
      "FacetsResponse": {
        "type": "object",
        "description": "Distinct filter values across all runs, ignoring the active filters.",
        "required": [
          "platforms",
          "device_classes",
          "device_serials",
          "models",
          "git_branches",
          "sumd_driver_versions",
          "bsp_versions",
          "host_accelerators"
        ],
        "properties": {
          "platforms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Platform"
            }
          },
          "device_classes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DeviceClass"
            }
          },
          "device_serials": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "models": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResultModelRef"
            }
          },
          "git_branches": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sumd_driver_versions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "bsp_versions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "host_accelerators": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "MetricStatsResponse": {
        "type": "object",
        "description": "Median, minimum, maximum, and sample count of one throughput metric,\nin tokens per second, over succeeded runs.",
        "required": [
          "median",
          "min",
          "max",
          "n"
        ],
        "properties": {
          "median": {
            "type": "number",
            "format": "double",
            "description": "Median in tokens per second."
          },
          "min": {
            "type": "number",
            "format": "double",
            "description": "Minimum in tokens per second."
          },
          "max": {
            "type": "number",
            "format": "double",
            "description": "Maximum in tokens per second."
          },
          "n": {
            "type": "integer",
            "format": "int64",
            "description": "Number of succeeded runs the statistic was computed over.",
            "minimum": 0
          }
        }
      },
...
      "Platform": {
        "type": "string",
        "description": "The kind of host a run executed on. Selects which host-state snapshot a\nrun carries. See `specs/benchmark-schema` - \"Host state is captured as a\nplatform-specific immutable snapshot\".",
        "enum": [
          "android",
          "linux"
        ]
      },
...
      "ResultModelRef": {
        "type": "object",
        "description": "The model a configuration exercised.",
        "required": [
          "id",
          "original_name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "original_name": {
            "type": "string",
            "description": "The registered `.pte` file's name."
          }
        }
      },
      "ResultRowResponse": {
        "type": "object",
        "description": "One benchmark configuration and its statistics over all its runs.",
        "required": [
          "platform",
          "device_class",
          "device_serial",
          "model_asset",
          "git_commit_sha",
          "git_dirty",
          "prompt_sha256",
          "input_token_count",
          "total_runs",
          "not_succeeded",
          "correctness_failed",
          "throttled",
          "first_run_at",
          "last_run_at"
        ],
        "properties": {
          "platform": {
            "$ref": "#/components/schemas/Platform"
          },
          "device_class": {
            "$ref": "#/components/schemas/DeviceClass",
            "description": "`internal`: lab device. `external`: retail phone or Linux box."
          },
          "device_serial": {
            "type": "string",
            "description": "Android: device serial. Linux: hostname."
          },
          "device_model": {
            "type": [
              "string",
              "null"
            ],
            "description": "Product/model name as the host reports it. Null when not captured."
          },
          "model_asset": {
            "$ref": "#/components/schemas/ResultModelRef"
          },
          "git_commit_sha": {
            "type": "string"
          },
          "git_dirty": {
            "type": "boolean"
          },
          "sumd_driver_version": {
            "type": [
              "string",
              "null"
            ],
            "description": "Lab Android devices only; null elsewhere."
          },
          "bsp_version": {
            "type": [
              "string",
              "null"
            ],
            "description": "Lab Android devices only; null elsewhere."
          },
          "gpu_clock_mhz": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "GPU clock in MHz. Lab Android devices only; null elsewhere."
          },
          "mif_clock_mhz": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "MIF (memory interface) clock in MHz. Lab Android devices only; null\nelsewhere."
          },
          "int_clock_mhz": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "INT (interconnect) clock in MHz. Lab Android devices only; null\nelsewhere."
          },
          "host_accelerator": {
            "type": [
              "string",
              "null"
            ],
            "description": "The accelerator the runtime backend executed on. Null when not\ncaptured."
          },
          "prompt_sha256": {
            "type": "string",
            "description": "Lowercase 64-character hexadecimal SHA-256 digest of the prompt.",
            "example": "b5bb9d8014a0f9b1d61e21e796d78dccdf1352f23cd32812f4850b878ae4944"
          },
          "input_token_count": {
            "type": "integer",
            "format": "int64"
          },
          "git_branch": {
            "type": [
              "string",
              "null"
            ]
          },
          "git_commit_timestamp": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "git_commit_subject": {
            "type": [
              "string",
              "null"
            ]
          },
          "prefill": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/MetricStatsResponse",
                "description": "Prefill throughput statistics in tokens per second over succeeded\nruns. Null when no run of this configuration succeeded."
              }
            ]
          },
          "decode": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/MetricStatsResponse",
                "description": "Decode throughput statistics in tokens per second over succeeded\nruns that recorded a decode measurement. Null when none did."
              }
            ]
          },
          "total_runs": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "not_succeeded": {
            "type": "integer",
            "format": "int64",
            "description": "Runs whose exit status is not `succeeded`.",
            "minimum": 0
          },
          "correctness_failed": {
            "type": "integer",
            "format": "int64",
            "description": "Succeeded runs whose correctness result is `failed`.",
            "minimum": 0
          },
          "throttled": {
            "type": "integer",
            "format": "int64",
            "description": "Runs that reported thermal throttling.",
            "minimum": 0
          },
          "first_run_at": {
            "type": "string",
            "format": "date-time"
          },
          "last_run_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ResultsResponse": {
        "type": "object",
        "description": "Grouped results: one row per configuration, newest commit first.",
        "required": [
          "rows",
          "truncated",
          "facets"
        ],
        "properties": {
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResultRowResponse"
            }
          },
          "truncated": {
            "type": "boolean",
            "description": "True when more than the maximum number of configurations matched\nand only the first 500 are returned."
          },
          "facets": {
            "$ref": "#/components/schemas/FacetsResponse"
          }
        }
      },
...
    }
  },
...
}